Seperated out Partitioner struct into its own files
